🚀 chore(front-end): add redux dependencies and update App.tsx to use redux ✨ feat(front-end): add redux slice to handle user data The redux dependencies have been added to the package.json file. The App.tsx file has been updated to use redux. The userSlice has been added to handle user data. The setUsername action has been added to update the username in the store. The setUserDatas action has been added to update the user data in the store. The user data is now displayed in the App component and can be updated by clicking the "update user" button.
🗑️ chore(react.svg): delete react.svg file
The react.svg file has been deleted from the front-end/src/assets directory.

🎉 feat(front-end): add redux store and types
The commit adds support for Redux store and types. The store is configured using `@reduxjs/toolkit` and `redux-persist` to persist the user data. The `Provider` and `PersistGate` components are added to the `main.tsx` file to provide the store to the application. The `userSliceProps` and `userSliceState` types are added to the `types.tsx` file to define the structure of the user data.

🎉 feat(userSlice.tsx): add userSlice reducer
This commit adds a new userSlice reducer to the front-end application. The reducer contains two actions, setUsername and setUserDatas, which update the state of the userSliceProps. The initialState of the reducer contains default values for the userSliceProps.