🚚 chore(Logout.tsx): rename setNames to setUserInfos The setNames function in the userSlice has been renamed to setUserInfos to better reflect its purpose of setting all user information, not just the name.
🔧 refactor(SignInForm.tsx): add dispatch to setUserInfos action
The setUserInfos action is now dispatched with the user's first name, last name, email, and id. This allows the user's information to be stored in the Redux store, which can be accessed throughout the application.

🔨 refactor(userSlice.tsx): rename setNames to setUserInfos and update its payload
The setNames function has been renamed to setUserInfos to better reflect its purpose. The payload has been updated to include all user information, including id and email. This change improves the clarity of the code and makes it easier to understand what data is being set.