🎨 style(swagger.yaml): change quotes to double quotes and fix indentation This commit only changes the quotes used in the swagger.yaml file from single quotes to double quotes. It also fixes the indentation of the file to make it more readable.
🎨 style(swagger.yaml): fix indentation and quotes
The commit fixes the indentation and quotes in the swagger.yaml file. The indentation is now consistent and the quotes are now double quotes throughout the file.

🚀 chore(swagger.yaml): add CRUD operations for User Transaction API
This commit adds CRUD operations for the User Transaction API. The API now supports creating, reading, updating, and deleting user transactions. The API is secured with a Bearer JWT token.

✨ feat(swagger.yaml): add UserTransaction definition
The UserTransaction definition is added to the swagger.yaml file. It is an object that has three properties: amount, description, and type. The amount property is of type number, the description property is of type string, and the type property is of type string with an enum of credit and debit. This definition will be used to represent a user transaction in the API.