🚚 chore(swagger.yaml): move user transactions endpoints to /user/profile/transactions 📝 docs(swagger.yaml): update descriptions for user transactions endpoints The user transactions endpoints have been moved to /user/profile/transactions to better reflect the resource being accessed. The descriptions for the endpoints have been updated to reflect this change. The description for the /user/profile/transactions endpoint has been updated to remove the reference to "a" user transactions and the description for the /user/profile/transactions/{id} endpoint has been updated to clarify that it fetches a single user transaction.
🚀 feat(swagger.yaml): add Transaction Module and update UserTransaction schema
The Transaction Module has been added to the swagger.yaml file. The UserTransaction schema has been updated to UserTransactions to reflect the fact that it is now an array of transactions. The post method has been updated to create a new transaction, and the patch method has been updated to update a transaction. The delete method has been removed. The ApiResponse schema has been updated to ApiResponseSingleTransaction to reflect the fact that it is now a response for a single transaction. The 404 and 503 response codes have been added to the get method.

🔥 chore(swagger.yaml): remove unused post endpoint and update delete endpoint
The post endpoint for creating a user transaction has been removed as it is no longer needed. The delete endpoint has been updated to include a path parameter for the transaction ID that is to be deleted. This improves the clarity of the endpoint and makes it easier to use.

🚀 feat(swagger.yaml): add new response objects for transactions
This commit adds new response objects for transactions. ApiResponseTransactions and ApiResponseSingleTransaction are added to handle responses for multiple transactions and a single transaction respectively. UserTransactions and UserSingleTransaction are added to handle the request body for multiple transactions and a single transaction respectively. These new response objects will improve the API's ability to handle transactions.